refactor: renames dateUtils into date
