+++ frontend/amundsen_application/static/js/pages/TableDetailPage/index.tsx
@@ -55,7 +55,7 @@ import TagInput from 'features/Tags/TagInput';
 import LoadingSpinner from 'components/LoadingSpinner';
 
 import { logAction, logClick } from 'utils/analytics';
-import { formatDateTimeShort } from 'utils/dateUtils';
+import { formatDateTimeShort } from 'utils/date';
 import {
   buildTableKey,
   getLoggingParams,
