+++ frontend/amundsen_application/static/js/utils/stats.ts
@@ -1,4 +1,4 @@
-import { formatDate } from 'utils/dateUtils';
+import { formatDate } from 'utils/date';
 
 import { getUniqueValueStatTypeName } from 'config/config-utils';
